requirements.txt
Django==1.5.4
Pillow==2.2.2
django-blog-zinnia==0.13
django-mptt==0.6.0
django-tagging==0.3.1
django-xmlrpc==0.1.5
psycopg2==2.5.1
pyparsing==2.0.1
pytz==2013.8
six==1.4.1
wsgiref==0.1.2
